Renamed TestView and MonopolyGame
